Factor out column selector
